likes limitados e functions para posts publicos ou privados
