+++ src/pages/feed.js
@@ -1,5 +1,6 @@
 import Button from '../components/button.js';
 import Textarea from '../components/textarea.js';
+import { AddComment, PrivacyPost, EditPost, LikePost, DeletePost } from '../posts/functions.js';
 
 function logOut() {
   auth
@@ -23,6 +24,7 @@ function createPost() {
   const text = document.querySelector('.text-area').value;
   const post = {
     likes: 0,
+    user_likes: [],
     text,
     comments: [],
     user_name: firebase.auth().currentUser.displayName,
@@ -39,84 +41,6 @@ function createPost() {
       console.log('Não foi possível criar post.');
     });
   document.querySelector('.text-area').value = '';
-}
-
-function printComments(arr) {
-  let template = '';
-  arr.forEach((text) => {
-    template += `
-    <li class='comments-list'>${text.userName}<br>${text.newComment}</li>
-    `;
-  });
-  return template;
-}
-
-function addPost(post, postId) {
-  const LoggedUserID = firebase.auth().currentUser.uid;
-  const postTemplate = `
-      <li class='post' id="${postId}">
-        <p>Postado por ${post.user_name} em ${post.createdAt.toDate().toLocaleString('pt-BR').substr(0, 19)}</p>
-        <p class="post-text">${post.text}</p>
-        ${LoggedUserID === post.user_id ? '<div class="delete fa fa-trash"></div> <div><span class="edit-post fa fa-pencil"></span></div>' : ''}
-        <div class="edit-button"></div>
-        <div class="interaction-area">
-        <div class="like fa fa-heart"></div>
-        ${post.likes}
-        <span class='comment-icon fa fa-comments'></span>
-        <div class="comments">
-          <div class="comment-container"></div>
-          <ul class='comment-posts'>${printComments(post.comments)}</ul>
-        </div>
-      </li>
-      `;
-  return postTemplate;
-}
-
-function saveComment() {
-  const newComment = document.querySelector('.textarea-comment').value;
-  const datasetid = event.target.dataset.id;
-  db.collection('users').doc(auth.currentUser.uid).get().then((doc) => {
-    const userName = doc.data().name;
-    const docPost = db.collection('posts').doc(datasetid);
-    docPost.update({
-      comments: firebase.firestore.FieldValue.arrayUnion({
-        userName,
-        newComment,
-      }),
-    });
-  });
-}
-
-function cancelComment() {
-  const datasetid = event.target.dataset.id;
-  document.getElementById(datasetid).querySelector('.comment-container').innerHTML = '';
-}
-
-function addComment(postId) {
-  const commentArea = `
-    ${Textarea({
-    class: 'textarea-comment',
-    placeholder: 'Escreva um comentário',
-  })}
-    ${Button({
-    type: 'button',
-    class: 'btn',
-    id: 'btn-comment-post',
-    dataId: postId,
-    onclick: saveComment,
-    title: 'Postar',
-  })}
-    ${Button({
-    type: 'button',
-    class: 'btn',
-    id: 'btn-comment-cancel',
-    dataId: postId,
-    onclick: cancelComment,
-    title: 'Cancelar',
-  })}
-  `;
-  const createSection = document.getElementById(postId).querySelector('.comment-container');
-  createSection.innerHTML = `${commentArea}`;
 }
 
 function NewPostTemplate() {
@@ -147,77 +71,38 @@ function NewPostTemplate() {
   return template;
 }
 
-function save() {
-  const id = event.target.dataset.id;
-  const postText = document.getElementById(id).querySelector('.post-text');
-  const saveEdit = document.querySelector('.edit-textarea').value;
-  postText.innerHTML = `
-  <p class='post-text'>${saveEdit}</p>
-  `;
-  firebase.firestore().collection('posts').doc(id).update({
-    text: saveEdit,
+function printComments(arr) {
+  let template = '';
+  arr.forEach((text) => {
+    template += `
+    <li class='comments-list'>${text.userName}<br>${text.newComment}</li>
+    `;
   });
-  document.getElementById(id).querySelector('.edit-button').innerHTML = '';
+  return template;
 }
 
-function cancel() {
-  const id = event.target.dataset.id;
-  const postText = document.getElementById(id).querySelector('.post-text');
-  const text = postText.textContent;
-  postText.innerHTML = `
-  <p class='post-text'>${text}</p>
-  `;
-  document.getElementById(id).querySelector('.edit-button').innerHTML = '';
-}
-
-function editPost(postId) {
-  const id = postId;
-  const postText = document.getElementById(id).querySelector('.post-text');
-  const button = document.getElementById(id).querySelector('.edit-button');
-  const text = postText.textContent;
-  postText.innerHTML = `
-  ${Textarea({
-    class: 'edit-textarea',
-    id: 'edit-textarea',
-    placeholder: '',
-    value: text,
-  })}
-  `;
-  button.innerHTML = `
-    ${Button({
-    id: 'btn-save',
-    class: 'btn save-btn',
-    dataId: postId,
-    onclick: save,
-    title: 'Salvar',
-  })}
-    ${Button({
-    id: 'btn-cancel',
-    class: 'btn cancel-btn',
-    dataId: postId,
-    onclick: cancel,
-    title: 'Cancelar',
-  })}
-  `;
-}
-
-function deletePost(postId) {
-  if (!confirm('Tem certeza que deseja excluir essa publicação?')) return;
-  const postsCollection = firebase.firestore().collection('posts');
-  postsCollection.doc(postId).delete().then(() => {
-  })
-    .catch((error) => {
-      console.log(error);
-    });
-}
-
-async function likePost(postId) {
-  const postsCollection = firebase.firestore().collection('posts');
-  const actualPost = await postsCollection.doc(postId).get();
-  postsCollection.doc(postId).set({
-    ...actualPost.data(),
-    likes: ++actualPost.data().likes,
-  });
+function addPost(post, postId) {
+  const LoggedUserID = firebase.auth().currentUser.uid;
+  const postTemplate = `
+      <li class='post' id="${postId}">
+        <p>Postado por ${post.user_name} em ${post.createdAt.toDate().toLocaleString('pt-BR').substr(0, 19)}</p>
+        <p class="post-text">${post.text}</p>
+        ${LoggedUserID === post.user_id ? '<div class="delete fa fa-trash"></div> <div><span class="edit-post fa fa-pencil"></span></div>' : ''}
+        <div class="edit-button"></div>
+        <div class="interaction-area">
+        <div class="like fa fa-heart"></div>
+        ${post.likes}
+        <div class='comment-icon fa fa-comments'></div>
+        <!----------------------------------------------------Monica mexeu aqui-->
+        ${LoggedUserID === post.user_id ? '<select class="privacy"><option value="fa fa-globe">&#xf0ac; Público</option><option value="fa fa-lock">&#xf023; Privado</option></select>' : ''}
+        <!----------------------------------------------------Monica mexeu aqui-->
+        <div class="comments">
+          <div class="comment-container"></div>
+          <ul class='comment-posts'>${printComments(post.comments)}</ul>
+        </div>
+      </li>
+      `;
+  return postTemplate;
 }
 
 function loadPosts() {
@@ -230,24 +115,33 @@ function loadPosts() {
     });
     document.querySelectorAll('.delete').forEach((btn) => {
       btn.addEventListener('click', (event) => {
-        deletePost(event.target.parentNode.getAttribute('id'));
+        DeletePost(event.target.parentNode.getAttribute('id'));
       });
     });
     document.querySelectorAll('.like').forEach((btn) => {
       btn.addEventListener('click', (event) => {
-        likePost(event.target.parentNode.parentNode.getAttribute('id'));
+        LikePost(event.target.parentNode.parentNode.getAttribute('id'));
       });
     });
     document.querySelectorAll('.edit-post').forEach((btn) => {
       btn.addEventListener('click', (event) => {
-        editPost(event.target.parentNode.parentNode.getAttribute('id'));
+        EditPost(event.target.parentNode.parentNode.getAttribute('id'));
       });
     });
     document.querySelectorAll('.comment-icon').forEach((icon) => {
       icon.addEventListener('click', (event) => {
-        addComment(event.target.parentNode.parentNode.getAttribute('id'));
+        AddComment(event.target.parentNode.parentNode.getAttribute('id'));
       });
     });
+    ///////////////////////////////////Monica mexeu aqui
+    document.querySelectorAll('.privacy').forEach((selection) => {
+      selection.addEventListener('change', (event) => {
+        let targetOption = document.querySelector('.privacy').options[document.querySelector('.privacy').selectedIndex].value
+        //console.log(targetOption)
+        PrivacyPost(event.target.parentNode.parentNode.getAttribute('id'), targetOption);
+      });
+    });
+    ///////////////////////////////////Monica mexeu aqui
   });
 }
